26. Setup Hearing Sense + ( UPawnNoiseEmitterComponent ) + ( Seen is more important to AI than Noise!)
Setup the "Hear Noise" delegate and boud it:
- DELEGATE: PawnSensingComp->OnHearNoise.AddDynamic(this, &AAIGuard::HandleOnHearNoise)
- BIND: void AAIGuard::HandleOnHearNoise(APawn* Instigator, const FVector& Location, float Volume)

HandleOnHearNoise():
- DrawDebugSphere(GetWorld(), Location, 32.0f, 12, FColor::Red, false, 10.0f);    //Instead of using the heard pawn (instigator), we just use the location the delegate returned :D)

In order for us to hear the player character, we need to add a noise emmitter to the player character.  (If we call the UE4 function 'PawnMakeNoise()', but we don't have an emitter component on the pawn we want to make the noise, it will do nothing.
- NoiseEmitterComp = CreateDefaultSubobject<UPawnNoiseEmitterComponent>(TEXT("NoiseEmitterComp"));   //also created the type, forward declarations and includes.

- NOTE!!! While you are seen, it will no longer execute the noise event.