38. Network the Game State Part 1
Because we added (Role == ROLE_Authority) to the code that sets the bIsCarryingObjective to true, the variable is only set on the server side, and can never be true for clients. To fix this, we will set this var to be Replicated.

- On Character, we made the bIsCarryingObjective replicated.
- Added the override function to set the rules (GetLifetimeReplicatedProps)
- defined the override function, added Super, AND Specified the var we want to repliate + Update the class.

Note: We could have used
DOREPLIFETIME_CONDITION(AAIGuard, GuardState, COND_OwnerOnly);  //We could have used this. COND_OwnerOnly adds the replication to the owner. This would optimize the code.

Note:
Remember that the game mode doesn't run on clients. It only exists on the server. Thus, added pointer protection to the Extraction Zone (line 44)