39. Network the Game State Part 2 + ( GameState C++ Class ) + ( GetPawnIterator() )
The Problem: GameMode it not something that can be replicated.
The Solution: Leverage GameStates :D

- We created our own C++ class of Gamestate,
- We setup a UFUNCTION(NetMulticast, Reliable) function called MultiCastOnMissionCompleted(), with the same inputs as OnMissionCompleted(), so that we can pass the same variables.
- We then impletented this function with MultiCastOnMissionCompleted_Implementation(), because its a NetMulticast. We also added a _Validate(), because its a Reliable.
- In MultiCastOnMissionCompleted_Implementation(), we used the following to loop through pawns, and check if they are controlled locally, to disable inputs:

UWorld::GetPawnIterator()                             - This can be used to cycle through all the pawns in the world with a for loop.
APawn::IsLocallyControlled()                        - This checks if a pawn is controlled by a player.

- Then back in the GameMode Class, we set our default GameState class to our new C++ version, with :
GameStateClass = AFPSGameState::StaticClass();                            -  Set our Default setting for Game State (This could also be set in the editor of our game mode).

- Lastly, we removed the disable input call from the function, and  called the the game state function to do this instead. We used:
AGameModeBase:: GetGameState()          - This gets the game state from the game mode.